// _js/snippets/product-options/app/modules/as-swatch/style.module.scss.d.ts
/* DO NOT EDIT: This file is auto-generated and will be overwritten. [build-fingerprint:codegen] */
export type Styles = {
  "OptionSwatches": "OptionSwatches",
  "OptionSwatches-badge": "OptionSwatches-badge",
  "OptionSwatches-indicator": "OptionSwatches-indicator",
  "OptionSwatches-indicatorImage": "OptionSwatches-indicatorImage",
  "OptionSwatches-item": "OptionSwatches-item",
  "OptionSwatches-item--outOfStock": "OptionSwatches-item--outOfStock",
  "OptionSwatches-item--selected": "OptionSwatches-item--selected",
  "OptionSwatches-item--unavailable": "OptionSwatches-item--unavailable",
  "OptionSwatches-itemInput": "OptionSwatches-itemInput",
  "OptionSwatches-label": "OptionSwatches-label",
  "OptionSwatches-list": "OptionSwatches-list",
  "OptionSwatches-title": "OptionSwatches-title",
  "OptionSwatches-titleDescription": "OptionSwatches-titleDescription",
  "OptionSwatches-titleSelection": "OptionSwatches-titleSelection"
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles